+++ test/HspecMain.hs
@@ -11,6 +11,7 @@ import qualified Spec.UnicodeHandling
 import qualified Spec.GoldenWire
 import qualified Spec.ModernEra
 import qualified Spec.ProtocolVersionNegotiation
+import qualified Spec.TemplatesCompletions
 import qualified Spec.ToolCallParsing
 import qualified Spec.TypedArgs
 
@@ -25,5 +26,6 @@ main = hspec $ do
     Spec.GoldenWire.spec
     Spec.ModernEra.spec
     Spec.ProtocolVersionNegotiation.spec
+    Spec.TemplatesCompletions.spec
     Spec.ToolCallParsing.spec
     Spec.TypedArgs.spec